add Ctrl+N as a shortcut to add item in trigger editor (#6907)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
This adds a keyboard shortcut to the Add Item button so that when your
focus is in the tree widget, you press Ctrl+N and it does the same thing
as clicking that button. Ctrl+Shift+N does the same for groups.
Also changed from current behavior of the button, made it so that adding
an item puts focus into name box for item and selects it for typing a
replacement.
#### Motivation for adding to Mudlet
Avoid having to use tab to navigate to the item in the bar.
#### Other info (issues closed, discussion etc)
Discussion today on accessibility channel on Discord.
`QKeySequence::New` is `Ctrl+N` on all the platforms
https://doc.qt.io/qt-5/qkeysequence.html
`Ctrl+Shift+N` is one I made up.