Improve: save and restore multi-view mode between Mudlet application runs
This will, I think, close #6917.

Although the situation as described in that issue seems to be the opposite
to how the OP reported it, it is true that the multi-view mode is NOT saved
between Mudlet runs. This PR fixes that so that the multi-view mode **is**
now saved between runs so that if it is set when Mudlet closes it will then
be applied in a future run once a second profile is opened.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>