Fix: ensure all TConsole instances have a name
This is a more comprehensive fix that also obsoletes #6926 which should be
closed to defer to this one. As per that Pull-request it seems that #6232  -
(commit 3c7593ab) removed a default value for
`(QString) TConsole::mConsoleName` which was `main` - I guess this was not
spotted sooner because much of our code assumes that an unnamed `TConsole`
identifier is the `TMainConsole` instance for a profile!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>