Fix: crash on moving rooms in new map in new profile
In fact this ought to be happening in other cases as well when using the
"Move to position" 2D Mapper context menu option. It is a result of setting
the `Qt::WA_DeleteOnClose` attribute on the `QDialog` used to get the
user's input and using `QDialog::exect()` in order to "run" the dialog. As
a result it deletes itself when the user clicks the "Okay" button but
Mudlet has been trying (and crashing when it cannot) to read the text
contents of the `QLineEdit`s that the user has typed their input into
after that point!

The fix is to NOT set that attribute (ensuring that the dialog is
destroyed AFTER the details have been gathered - or even if they are not
because "Cancel" has been clicked instead).

This should close #6932. As a check I have reviewed all the other places
where `QDialog::exec()` is used and I have determined to my satisfaction
that there are no other cases in our code where this could also happen.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>