Improve: remove speedWalkDir normal exit directions from translation system
Whilst describing how the SpeedWalking code operated last night on Discord
in the **#help** channel:
https://www.discord.com/channels/283581582550237184/283582068334526464/1142680496254484561
I realised that we had been passing the normal exit directions through the
translation system for the end-users GUI language. This is wrong because
there is not necessarily a match between the language the end-user prefers
and that used by any MUD they might play. As such it means that the normal
exit "abbreviations" that the speedwalking code uses to fill in the
`speedWalkDir` table change depending on which GUI language is selected and
this means that the "mapper" package for any MUD has to know that and then
translate them to the language that the MUD uses - as they need not be the
same.

This PR provides a means to disable doing the (unhelpful) translation to
user's locale and will do that by default from now on. However should it
be needed (because the user or "mapper" package provider has added code) to
do this extra conversion it can be re-enabled by a checkbox in the "Special
options" tab of the preferences or by using the
`"translateNormalExitsInSpeedWalkDirTable"` key with the Lua `setConfig(...)`/
`getConfig(...)` pair of functions -  the presence of that key in those
functions can be used to detect that this capability is present in a
particular Mudlet build.

Examining the various `translations/translated/mudlet_xx_YY.ts` files I
have determined that this bogus translation in the user's GUI language
happens for the following language/country codes (the code that is used
for the others and for all cases when this PR is place without the bogus
translation being done AND the "direction codes" also used within the Lua
sub-system are also shown):

|code|Language (Country)|North|North-east|North-west|East|West|South|South-east|South-west|Up|Down|In|out|
|-|-|-|-|-|-|-|-|-|-|-|-|-|-|
||Direction code|`1`|`2`|`3`|`4`|`5`|`6`|`7`|`8`|`9`|`10`|`11`|`12`|
||All others (or after this PR)|`n`|`ne`|`nw`|`e`|`w`|`s`|`se`|`sw`|`up`|`down`|`in`|`out`|
|ar_SA|Arabic (Saudi Arabia)|`n`(?)|`ne`(?)|`nw`(?)|`e`(?)|`ص`|`s`(?)|`se`(?)|`sw`(?)|`فوق`|`تحت`|`للداخل`|`للخارج`
|de_DE|German (Germany)|`n`|`no`|`nw`|`o`|`w`|`s`|`so`|`sw`|`oben`|`unten`|`rein`|`raus`|
|es_ES|Spanish (Spain)|`n`|`ne`|`no`|`e`|`o`|`s`|`se`|`so`|`arriba`|`abajo`|`adentro`|`afuera`|
|fr_FR|French (France)|`n`|`ne`|`no`|`e`|`o`|`s`|`se`|`so`|`haut`|`bas`|`entrer`|`sortir`|
|it_IT|Italian (Italy)|`n`|`ne`|`no`|`e`|`o`|`s`|`se`|`so`|`alto`|`basso`|`dentro`|`fuori`|
|nl_NL|Dutch (Netherlands)|`n`|`no`|`nw`|`o`|`w`|`s`(?)|`zo`|`zw`|`omhoog`|`omlaag`|`in`|`uit`|
|pl_PL|Polish (Poland)|`n`|`ne`|`nw`|`e`|`w`|`s`|`se`|`sw`|`u`|`d`|`do środka`|`na zewnątrz`|
|pt_BR|Portuguese (Brazil)|`n`|`ne`(?)|`nw`(?)|`e`(?)|`w`(?)|`s`|`se`(?)|`sw`(?)|`cima`|`baixo`|`dentro`|`fora`|
|pt_PT|Portuguese (Portugal)|`n`|`ne`(?)|`no`|`e`(?)|`o`|`s`|`se`(?)|`so`|`cima`|`baixo`|`dentro`|`fora`|
|ru_RU|Russian (Russia)|`n`|`ne`|`nw`|`e`|`w`|`s`|`se`|`sw`|`up`|`down`|`in`|`out`|
|tr_TR|Turkish (Türkiye)|`k`|`kd`|`kb`|`d`|`b`|`g`|`gd`|`gb`|`y`|`a`|`i`|`d`(!)|
|zh_TW|Chinese (Traditional)|`n`|`ne`|`nw`|`e`|`w`|`s`|`se`|`sw`|`上`|`下`|`入口`|`出口`|

Note that there are some suspect strings in there (?) and also one case
where the same code is produced for two different exit directions - which
is not just suspect, but probably wrong!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>