Improve: Add Mute Capability for Media (#6961)
#### Brief overview of PR changes/additions
To encourage more people to use the media functions in Mudlet (i.e.,
playSound, MCMP, MSP), functionality to easily mute media in Mudlet is
added. A Mute button is added to the toolbar for quick access to mute
all media, or use the embedded actions in the Mute button or menu to
mute or unmute the 3 classes of media protocols:

1. Mudlet API (Scripts, Triggers, etc.)
2. Game MCMP (Mud Client Media Protocol sent via GMCP)
3. Game MSP (Mud Sound Protocol sent via MSP or MXP)

Shortcut key to Mute/Unmute All Media is `Alt + K` or for macOS `Command
+ K`.

#### Motivation for adding to Mudlet
Player feedback and discussion on the Mudlet Discord.

#### Other info (issues closed, discussion etc)
A test plan is added below, but if your MUD uses sound and mute and
unmute works, that is a great test and is good testing feedback. Scroll
to the `Test Versions` section below, pick your platform, install, and
connect to your favorite MUD with sound to try it out.

Contributed by Tamarindo@StickMUD.  

Icons sourced from
[opengameart.org](https://opengameart.org/content/play-pause-mute-and-unmute-buttons)
with a [CC0 1.0](https://creativecommons.org/publicdomain/zero/1.0/)
license.

**Release Notes**
Added easy access to mute and unmute sound via the toolbar, menu, or via
shortcut.

**Testing**
1. Use this command to download and play music via the Lua API:
```lua
lua playMusicFile({url = "https://www.iamtalon.me/", name = "spacemusic.mp3"})
```
2. Use `Alt + K` or for macOS `Command + K` to Mute and Unmute the
music.
3. Use the `Mute` or `Unmute` button on the toolbar to Mute and Unmute
the music.
4. Use the `Mute All Media` action under Options in the menu to Mute and
Unmute the music.
5. While unmuted, use the `Mute Mudlet API (Triggers, Scripts, etc.)`
action under the `Mute` button on the toolbar to Mute and Unmute the
music.
6. Repeat the above with other protocols and/or across multiple profiles
as desired.