Fix: improve symbols on map pan buttons
On some MacOS users machines the map pan buttons are not included in the
font that is being used to render the mapper controls. One user has
suggested that we try some arrow symbols instead and I tried that already
in PR #6981. However those did seem a bit thin and I have an alternative
that I wish to test here.

This PR converts all 6 of the pan control buttons into `QToolButton`s
which have the feature that Qt can itself generate triangular "Arrow" icons
for - which *might* avoid the "emojification" that was taking place in the
current release code for some OS/Font combinations. The two Z direction
buttons do not have such arrows and retain their original "+" or "-" text.

This may close #5454.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>