Fix: use different symbols on map pan buttons
On some MacOS users machines the map pan buttons are not included in the
font that is being used to render the mapper controls. One user has
suggested that we try some arrow symbols instead. This PR changes them
as follows:
* from '⯅' U+2BC5 {BLACK MEDIUM UP-POINTING TRIANGLE CENTRED}
  to '↑' U+2191 {UPWARDS ARROW}
* from '⯆' U+2BC6 {BLACK MEDIUM DOWN-POINTING TRIANGLE CENTRED}
  to '↓' U+2193 {DOWNWARDS ARROW}
* from '⯇' U+2BC7 {BLACK MEDIUM LEFT-POINTING TRIANGLE CENTRED}
  to '←' U+2190 {LEFTWARDS ARROW}
* from '⯈' U+2BC8 {BLACK MEDIUM RIGHT-POINTING TRIANGLE CENTRED}
  to '→' U+2192 {RIGHTWARDS ARROW}

Another set was proposed by a different person but on my Windows machine
the left and right triangles got "emojified" which is what we are trying
to avoid. Ideally the chosen glyphs should not change like that for any of
of our three main supported OSes.

This may close #5454.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>