Fix: update Windows CI build of libzip
For a week or so these have been failing - supposedly because of a
certification problem when trying to download the source files to build
`libzip` however it seems that upstream have release a new version - which
according to https://libzip.org/packages/ has already made it's way into
MSYS2. This PR updates the version used for Windows CI builds from *1.7.3*
to **1.10.1** to see if that solves the problem.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>