Infrastructure: extend range of SGR codes captured
This is an initial step that I can isolate into a PR for submission
independently so as to keep the "chunk" size smaller and easier to digest
for reviewing. It detects and records some more SGR codes than we currently
do, specifically:
* 5 - slow, less than 150 times a minute, blinking
* 6 - fast, at least 150 times a minute, blinking
* 8 - concealed ON
* 10 - default font
* 11 - alternate font 1
* 12 - alternate font 2
* 13 - alternate font 3
* 14 - alternate font 4
* 15 - alternate font 5
* 16 - alternate font 6
* 17 - alternate font 7
* 18 - alternate font 8
* 19 - alternate font 9
* 25 - blink OFF
* 28 - concealed OFF

We already handle the two types of blinking by hard-coding them to italics,
this PR replicates that but at the point where the text is painted rather
then when the SGR parameters are parsed.

I already have some prototype code that provides support for painting in
different fonts but it needs a little more work to complete and I will
submit it in a following PR that will need this one to have been merged
first.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>