Fix: invalid grey QColor generated for grey-scale part of color triggers
There is supposed to be a label that demonstrates the grey value selected
when the "More colors" button is clicked for a color trigger item and the
24 position slider is adjusted. This color is also supposed to be replaced
in the foreground or background colour set for the color trigger and is
recorded as part of the details for a color trigger.

However, a defect in #5328 seems to have been based on the "slider-value"
being passed to `(void) dlgColorTrigger::slot_grayColorChanged(int)` being
in the range 232 to 255 rather than the 0 to 23 that it is. The result is
that the QColor generated from that grey scale part of the ANSI 256 color
range is wrong (invalid). This at least means that the grey color chosen
does not show up in the editor - it also may mean that such color triggers
may be defective and not work as expected.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>