Infrastructure: fix MacOS run time crashes due to missing libzstd
We have recently started to see crashes because MacOS builds do not appear
to be finding `libzstd.1.dylib` to dynamically link to at run-time as
needed by `libzip-5.dylib`, e.g. here:
https://github.com/Mudlet/Mudlet/pull/6982#issuecomment-1752202617

This tries to improve things by ensuring the library is now explicitly
listed as a package to have installed. Whether that is sufficient to fix
the problem needs to be tested. If it does not solve the problem we may
need to look to see how we specify what libraries/frameworks are required
in the bundled application that we provide on MacOS. I can see that (with
the `dmg2img` Linux utility) the library does seem to be absent from our
product currently (the recent PR test build that is mentioned about) but
**it is present, (as the `mudlet.app/Contents/Frameworks/libzstd.1.dylib`
file) in our most recent *Release* version...**

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>