Improve: add faint support and separate bold from colours
This disconnects the link that Mudlet had between ANSI colour numbers 8 to
15 (in the 16 and 256 colour ranges) from emboldening (increasing the
"weight") of the Font used.

It also adds support for the ANSI Escape SGR sequence "2" option for
"faint" (a.k.a. "dim") which provides two additional "weights" that can be
combined with the bold effect.

So as to show the effects four `QLineEdit`s and `QLabel`s have been added
beneath the font selection controls so that if the user types some text
into the first (normal) one the text is reproduced in all of them but with
the different combinations applied in each different box - there is also
some placeholder text in each of them which will still partially show the
effects even if nothing is entered into the first one.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>