Fix: selection of TTreeWidget items
#### Brief overview of PR changes/additions
Reverts and improves upon "improve Trigger Editor so pressing Enter would
activate and deactivate items (#6829)" (commit
cea6e429930a2fd6fe6bae9bb5dd4b9cc5bac702.) - so that double-clicking
will (again) toggle the item's activation state. To reproduce the desired
"Pressing Enter would activate and deactivate items" it adds a
`keyPressEvent` handler to the `TTreeWidget` class that responds only
to the press of the "Enter" or numeric keypad "Return" keys to raise a
new `enterOrReturnPressed` signal that can be used to also trigger the
same slot that the restored `QTreeWidget::itemDoubleClicked` signal does.

#### Motivation for adding to Mudlet
As per https://github.com/Mudlet/Mudlet/pull/6829#issuecomment-1639403026
#6829 makes the editor almost unusable for visually **un**impaired users
who wish to select items by clicking on them in the `TTreeWidget` as that
clicking to select them also toggles their state. This PR should rectify
that problem that I (@SlySven) have experienced along with the person
making that comment.

#### Other info (issues closed, discussion etc)
As a connected fix it also prevents the `QTreeWidget::itemClicked` signal
from having any effect on the second occasion where it is raised after
the `QTreeWidget::itemDoubleClicked` one is - this has been problematic
in the past and whilst those issues have been resolved it does not make
sense to run the connected slot functions a second time - this should make
the editor a tiny bit more responsive.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>