Revise: refactor CI/travis.osx.after_success.sh
This is so that there is less branching depending on whether it is a PTB or
release build - only one test rather than lots. This is so that it will be
easier to split the file up into two pieces - one for assembling the
package and one for signing it and uploading to our site.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>