Fix: improve symbols on map pan buttons (#6982)
#### Brief overview of PR changes/additions
This PR converts all 6 of the pan control buttons into `QToolButton`s
which have the feature that Qt can itself generate triangular "Arrow"
icons for - which *might* avoid the "emojification" that was taking
place in the current release code for some OS/Font combinations. The two
Z direction buttons do not have such arrows and retain their original
"+" or "-" text.

#### Motivation for adding to Mudlet
On some MacOS users machines the map pan buttons are not included in the
font that is being used to render the mapper controls. One user has
suggested that we try some arrow symbols instead and I tried that
already in PR #6981. However those did seem a bit thin and I have an
alternative that I wish to test here.

#### Other info (issues closed, discussion etc)
This may close #5454.

Only one of this and #6981 should be needed. :grinning:

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>