Improve: add BoldIsBright setting
Restore some of the removed code that allowed the Bold `<SGR>1m` code to combine with
and select colors 8 to 15 when the basic `<SGR>30m` to `<SGR>37m` codes for colors 0
to 7 to specify a foreground color. This is a known "hack" that some terminals as
well as MUD servers use to get sixteen colors. As it interferes with other, more
advances ways of selecting not just 16 but also 256 or even 16M colors this behaviour
is now made optional with a setting I have called "Bold is bright" which defaults to
"ON" to reproduce, *mostly* Mudlet's previous behaviour for foreground (but not
background) colors. However unless the user is playing a MUD that does use only 16
colors and does it this way they probably will want to switch this option OFF in
practice.

A further commit to follow is intended to make it easier for the end-user to see
what types of color selection codes the Game Server is sending so as to guide them
in that choice.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>