Fix: generic_mapper Window Shown config not updating (#6958)
#### Brief overview of PR changes/additions
When the value of a map_window sub_key is false, current behavior will
echo Unknown map config to the window and not change the value. This
change will check to see if the sub_key is nil instead, so that if the
value is the boolean false the value can be updated properly.

#### Motivation for adding to Mudlet
I was working on a GUI, and trying to set the value explicitly, and it
was reporting Unknown map config.

#### Other info (issues closed, discussion etc)
After determining this on my own, found that an issue had been opened
and come to the same conclusion, and this should close #6319.