Fix LDoc comment causing automated ldoc action to fail (#6977)
#### Brief overview of PR changes/additions
The automated action of building the LDocs hasn't been working for some
months due to an errant LDoc comment. See

https://github.com/Mudlet/Mudlet/actions/runs/6268731308/job/17024132323
for failure log. This PR fixes that and some improvements to the code
documentation.

- Fix all warnings
- Add a few new comments
- Add links to Mudlet Manual where appropriate
- Change some tags to LDoc style

#### Motivation for adding to Mudlet
Comment introduced in #6801 broke the automated build.

#### Other info (issues closed, discussion etc)
As above did some extra work on tidying up the docs. Still more to do
like some missing parameter and return comments, but this will fix the
automated action at least.