Improve: add code to detect specific SGR codes in incoming data
Monitors the incoming Server data and records the use of some `SGR`
formatting codes so that their use can be displayed in a section of the
"Display colours" tab of the profile preferences. As well as the type of
codes used to select the foreground and background colours it also monitors
the use of the Bold (1), Faint (2), Blinking (5), Fast Blinking (6) and
Concealed (8) codes. The first two, along with the color ones will help
the user to deduce how many colours the MUD Game Server might be using and
thus whether to enable or disabled the, now optional since the previous
commit, use of the Bold attribute to select the second set of eight
foreground colours of sixteen which some 16 colours MUDs and some terminals
use.

So as to reduce screen clutter the status is displayed on a set of disabled
(as they are not user adjustable) checkboxes that are all hidden until the
end-user requests them to be shown by clicking on the groupbox that
contains them and the button to reset the detection should they have just
told the Game Server that they want a different colour setting. The status
displays are updated in real-time and will get a checkmark as soon as the
corresponding code is detected on the incoming data. So as to avoid
confusion with anything that `feedTrigger(...)` might insert into the data
such data is disregarded - which does mean that testing cannot be carried
out using that function.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>