Improve: Port IRE mapping script functions to generic_mapper (#7020)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Port all functions as listed in issue. Aliases with a usage comment are
those I ported. Tried to keep alias name consistency. All new functions
are placed in same map script, but we should look at separating this out
as the script is becoming large and unwieldy.
Update version number to 2.1.0

#### Motivation for adding to Mudlet
/claim #5720

#### Other info (issues closed, discussion etc)
Also found issue in createMapLabel() and pushed to IRE repository. 
https://github.com/IRE-Mudlet-Mapping/ire-mapping-script/pull/191