Improve: add separate tooltip to insertPopup() and other functions (alternative) (#6946)
#### Brief overview of PR changes/additions
This is a replacement/alternative to PR #6925 which comprises a number
of changes that I have already squashed down to a single commit.

* Refactor/unify Lua (echo|insert|set)|(Link|Popup) functions

They have a lot of common code which can be refactored to make them
easier to comprehend.

As part of that we can also extend the ones involving popups (links with
context {right-click} menus) to accept an extra (initial) hint to use as
a tool-tip leveraging existing code put in to support the same from the
MXP `<SEND menu ...>`-tag protocol handler which was put in by PR #6877.

* Revise: inline (move to header) some simple `TLinkStore` class
methods.

* Revise: remove some bogus, unreachable, `return`s in TLuaInterpreter
class.

* Revise: use a better method to identify popup menu items

Previously they were keyed on the text shown for the items but this
breaks should (though it is admittedly unlikely) the same text be used
on more than one item. This could be needed for other developments which
might cause additional dummy/spacer/disabled entries to be present. Now
we store the index into the stored commands/lua functions `QMap` as a
data item that can be retrieved and used when a context menu action is
clicked upon.

* Revise: convert some implicit to explicit casts to remove some
warnings

* Revise: make popup actions without a command/function be disabled

* Revise: make popup items with both empty commands AND hints into
separators

* Revise: add missing `/test` files to QMake project file so they show
up in (qmake using) `qtcreator` application.

#### Motivation for adding to Mudlet
To allow popup menus to have a fancy "rich-text" format in the same
manner as @eowmob 's PR allows for from the same thing in MXP `<SEND
menu...>` tag features. He may not have realised the UI capabilities
that this might unlock, in that he may only have been thinking of pure
text to show when the mouse is hovered over such popups but by doing it
via a single (extra) hint/tooltip he actually made it easier to do than
my original attempt in #6925 did.

#### Other info (issues closed, discussion etc)
As a side effect this will also close an unrecorded issue I mentioned in
Discord about needing to provide the "main" window as a first argument
to `ìnsertPopup(...)`/`echoPopup(...)` if the last optional `(bool)
useCurrentFormat` argument is provided.

Additionally it is now possible to insert separators and disabled items
into the popups which could be useful to some end-users.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>