Improve: add ancestors functions (#6726)
Provides two Lua functions:
* `isAncestorsActive(itemID, "type")` - returns a boolean which is
`true` if all (if any) of the parents of the given item are set as
active and `false` if any are not. It makes use of an internal template
function that already exists but makes it available to the Lua
sub-system. If the item does not have any ancestors this function will
return `true`.
* `ancestors(itemID, "type")` - returns a table containing a sub-table
for each successively distance ancestor (if any) of the given item;
within each sub-table are details of the ancestor, specifically:
  * its ID as a number
  * its name as a string
  * whether it is active as a boolean
* its "node" (type), one of "item", "group" (folder) or "package"
(module)

***ALSO: adds a third, optional, boolean argument to `isActive(...)`
which if provided and `true` will only includes items in the count whose
parents are all enabled/active - if omitted it will be treated as if it
is `false` so that it behaves as the current version. Script-writers
should check for the presence of one of the other two functions to
determine if this third argument is handled.***

The first function will be enough to close #6724, specifically when it
is combined with `isActive(...)` as follows, it will enable a script to
determine whether an item (in this example a trigger with an ID number
stored as `myTriggerID`) will operate or not:
```lua
    if isAncestorsActive(myTriggerID, "trigger") and isActive(myTriggerID, "trigger") > 0 then
       -- Okay the aforesaid trigger will run!
    end
```

Note that these two new functions ONLY work with the ID number of an
item - this is because a unique identifier is required and Mudlet allows
for duplicate names so that cannot be used to reliably identify an item.
It had now been made a bit easier to determine the ID number of an
existing item the `findItems("name", "type")` function that has also been
implemented as#6742.

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>