Fix: Resolve raw telnet Client.GUI error (#7034)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Resolves an error that appears in the log for games that use the
non-JSON way to load Client.GUI from a server. The approach taken is to
skip setting a value in the GMCP table in this case where raw telnet is
used and then we avoid the error.

#### Motivation for adding to Mudlet
Issue registered in Mudlet

#### Other info (issues closed, discussion etc)

Closes #6739 

**Testing**
Create an account on Clessidra and when the GUI loads there will be no
error message in the log vs. the current production release of Mudlet
(4.17.2).

Co-authored-by: Michael Conley <mconley@michaels-mbp.lan>