Improve: add support for Telnet AreYouThere command
Following a discussion in the "#coding" channel in the MUD guild on Discord:
https://canary.discord.com/channels/279748146316312576/359062772337606656/1187507352346304583
I pointed out that Mudlet does not respond to the AYT command. This PR
fixes that so that it immediately responds (with one of a number of random
text phrases) when it gets that command - which should fulfil the
requirements for that.

It also increases the number of Telnet commands that get reported (to all
of them) in debug output - including the AYT one. To see them it will be
necessary to changed the `DEBUG_TELNET` #define from the `1` it had been
left as (perhaps it should have been left undefined after #2236 was merged)
to `2`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>