Improve: Add default template for new package in package exporter (#7041)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Add default template in markdown. Change textEdit_description behavior
to not clear on first click by assigning the markdown text on
initialisation.

#### Motivation for adding to Mudlet
/claim #5079 

#### Other info (issues closed, discussion etc)
Closes #5079.  

We should possibly look at making textEdit_description larger to
encourage the coder to make use of the space and help visually see the
markdown. Perhaps make it full window and using a two column display
with the editor using the entire right-side column.

![Peek 2023-12-21
15-52](https://github.com/Mudlet/Mudlet/assets/136661366/c6235848-bdb9-4f16-b572-2d3f72fa7db8)