fix: ensure that single clicks select without activating items (#7028)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Declare that single clicks should select but not not activate items in
the editor list.

#### Motivation for adding to Mudlet
#6829 made things feel broken on Linux using KDE specifically using the
single-click to activate item mode.

There is `QStyle::SH_ItemView_ActivateItemOnSingleClick` which is taken
from system but application can override it instead. For most setups it
gets 0, and for some setups it gets 1. This change should make it do 0
regardless.

Most should see absolutely no change with this.

The odd setups I'm aware of are that of [eternaleye with unknown
OS](https://github.com/Mudlet/Mudlet/pull/6829#issuecomment-1639403026)
and [SlySven with
FreeBSD](https://github.com/Mudlet/Mudlet/pull/7011#issuecomment-1848681252).

#### Other info (issues closed, discussion etc)