Add: Map mouse click event (sysMapWindowMousePressEvent) (#6962)
Add sysMapWindowMousePressEvent to know when a user clicked the map
window.

<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions

sysWindowMousePressEvent only works on the main window. This adds mouse
press events
to the mapper window.

#### Motivation for adding to Mudlet

To know _when_ a user has clicked the window, e.g. when a room is
selected.

#### Other info (issues closed, discussion etc)

Haven't added double clicks or release events. I couldn't think of a use
case. This is
also only the 2D mapper.