Improve: add findItems function (#6742)
#### Brief overview of PR changes/additions
This function is of the form:
`findItems("name", "type"[, exact[, case sensitive]])` with the last two
arguments defaulting to `true` if not given. `type` is one of "alias",
"button", "keybind", "script", "timer" or "trigger".

This returns a list of all the IDs for the type of item whose name
matches the given "name". If the exact argument is `false` the given
string causes any item whose name has it as a substring to be returned
in the results. As a side effect if an empty string is supplied then all
extant items of the specified type will be returned - which might be
useful in some situations.

#### Motivation for adding to Mudlet
The need for this function was something I posited at the end of:
https://github.com/Mudlet/Mudlet/pull/6726#issue-1651092803

#### Other info (issues closed, discussion etc)
A third PR has also been merged (#6743) that will show the ID
numbers in the editor...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>