Revert: undo prior commit that made an absolute move the default
This is because, if that change was also made to the putative `moveRoom(...)`
it would make that one almost functionally the same as
`setRoomCoordiantes(...)` and there is little need for two different
functions to do the same thing (except one might also move custom exit
lines). Reverting it also means that the default behaviour is closer to the
intended function which is so that map labels (and rooms and their exit
lines) can be "nudged" around the map without having to worry about their
existing coordinates.

Also make area specifier (1st argument) work for area names as well as area
id numbers.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>