Add: means to compose raw data for Lua Telnet testing and new protocols
#### Brief overview of PR changes/additions
This PR modifies the Lua API function `sendSocket(...)` and adds a new
function `feedTelnet(...)` so that there are means to create and issue data
containing the ASCII NULL byte value, and also adds a means to insert other
bytes that are useful when creating tests/protocol handlers for Telnet
sub-options.

#### Motivation for adding to Mudlet
The current implementation of `sendSocket(...)` and `feedTriggers(...)`
are not able to insert ASCII NULL characters into their data as it
immediately terminates the C-string that contains it. Furthermore the
latter does not insert its data into the "processing" early enough to
enable testing of the Telnet protocol/sub-option handling.

#### Other info (issues closed, discussion etc)
I discovered the need for something like this when I was reviewing #7058
and found that, even with the `\###` way (where ### is a three digit number
between `000` and `255`) of including unprintable characters it was not
possible to embed the `000` value as it gets converted into the ASCII
NULL character that terminates a "C" type string. Whilst constructing a way
to get around that - by using "tags" enclosed in `<`...`>` to represent
particular byte value I realised it would be useful to include tags for all
the ASCII control characters and also the Telnet control codes (with a `T_`)
prefix and also some ones (with a `O_` the letter O prefix) representing
some common or potentially useful Telnet sub-options - some of which Mudlet
supports already as well as some that might be done in the future. The
latter is important because without this capability it would not be
possible to generate required data streams in an add-on Lua package to send
to a MUD Game Server.

So that the table of tags used can be identified in the future calling
`feedTelnet("")` i.e. with an empty string, will obviously not generate
anything to send to the internal Mudlet telnet protocol handler, but will
prompt a simple version number stored in that function to be returned as
a second argument to the Lua caller.

This PR also adds a return value to `sendSocket(...)` which will return
true or false depending on whether it was actually possible to send the
data "up the socket" - which we had been determining but not returning
in the past. Also worth noting is that - to prevent confusing the internal
Telnet processing - `feedTelnet(...)` will refuse to send the generated
data into there unless the socket that would be also producing data to go
in is in an unconnected state - i.e. that the profile is loaded but not
connecting or connected to the Game Server.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>